Use define for all GDT entries.
Replace hardcoded values with define for all values in the GDT entries.
Each value is either the result of a simple define or of multiple
operations between define.
The define of the GDT address 'GDT_BASE' (now GDT_BASE_ADDRESS) has been moved to gdt.c as it is not used
in other files.

Signed-off-by: mrxx0 <chcoutur@student.42.fr>