Create and init the buddy allocator
Implement init functions for the buddy allocator, including:

kpm_init - which must be called before any other kpm function
and inits the buddy allocator itself in a reserved memory region.

kpm_enable/disable - Which can set memory regions as allocable or not.

For more details, please refer to the documentation, which will be made
available in the next commit

Signed-off-by: lfalkau <lfalkau@student.42.fr>