Implements the alloc and free functions
kpm_alloc:

The alloc function takes a size and will
determine the smallest order that can hold
this size.

It will then go through the bitmap of that order
to find a free block. If one is found, it will
allocate every child nodes on the first order,
fill the chunk structure, and update the parents
thanks to the kpm_update_orders function.

If none is found, it will decrement the order in which
it searches and try to find a smaller block. This operation
is repeated until the smallest order, at which point if no
block is found, it means that no more memory is available,
and the function returns an error code.

kpm_free:

The free function first checks that the provided chunk is aligned.
It then computes the frame index and the number of frames to free.
Now, for each frame, it will first check that the zone is enabled,
to avoid making a disabled zone available for future malloc calls
and free it if it is.
Finally, kpm_update_orders will be called to reflect the changes on
the whole bitmap.

Signed-off-by: lfalkau <lfalkau@student.42.fr>