Change the shell to a better one and fix strcmp.
Create ksh a simple shell with multiple arguments.
The shell call function found in an array call builtin in ksh.c and look
for a function with a name corresponmding to the first argument of the
command.
To execute a command, the shell translate the string entered by the user
to a array of string and call the correct function with the array of
string in argument of the function.
The shell is not perfectly working, all interaction with the screenbuf
execpt printing charactere doesn't work.

Fix strcmp: change `return s1 - s2;` to `return *s1 - *s2`.

Signed-off-by: glafond- <glafond-@student.42.fr>