Map the kernel at 0xc0000000 and turn on paging
_start call boot_init.
boot_init take three page at 0x1000, 0x2000 and 0x3000.
The first one is used as boot page directory and will be replace by a
new one when the virtual allocator will be ready.
The second one is used to map the boot page directory physical address
and all page table physical address to have the possibility to
modify/add page table or page entry when paging will be activated.
The third one is used a page table to map the first 4MB of memory.
boot_init put the boot page directory on cr3.

After boot_init the the paging is activated and an absolute jump is
done. Then the stack_top is loaded in the stack pointer and kernel_main
called.

Fix a wrong mask applied to the granularity of the gdt entries.

Remove temporary linker script and boot asm file.

Signed-off-by: xlmod <glafond-@student.42.fr>