Implement kvm_init, kvm_new_pagedirectory, kvm_map
kvm_new_pagedirectory tries to allocate one page with kpm_alloc and the
last page directory entry is mapped to the page directory itself and the
function returns the allocated page directory.

kvm_init calls kvm_new_pagedirectory then loads the new page directory in
cr3 register. Need to implement a better error handler.

kvm_map translate virtual address to a page directory entry index and a
page table entry index. As the last page directory entry maps itself, use
it to check if the page table, where the page will be mapped, is present.
Then map the page to the page table.
TODO: Create the page table if is not present and map it to the page
directory.

Create cr3 related functions:
- kvm_load_cr3
- kvm_get_cr3
- kvm_reload_cr3

Signed-off-by: xlmod <glafond-@student.42.fr>