Fix #48 and upgrade info buddy
In bitmap_ffu, use a bitmap_t(uint8_t) variable to get the inverted byte
and call __builtin_ffs.
The old way used a pointer and called `__builtin_ffs(~*cur)` that
casted the value in 32bit so the __builtin_ffs returned a index bigger
than the number of blocks.

Add to info buddy a optional arg to print the order.

Reset the color in alloc when allocation failed.

Check on size of alloc and free argument to return an error if negative.
Also check the address argument of free builtin to return an error if
not well formatted.

Signed-off-by: xlmod <glafond-@student.42.fr>
Co-authored-by: mrxx0 <chcoutur@student.42.fr>