Create some vmm structs and functions
Create vmzone struct and functions that describe a virtual memory zone
with a start and end pointer, some flags, a name and a list of backing
memory chunk.

Create mapping function to map physical address on virtual space.
Map_page and map_zone map on the loaded page directory and map_page_to
and map_zone_to map on a given page directory.
Quickmap_page function is used to map a physical page on the same
virtual zone (placed behind the stack), so each call overrides the
previous mapping.

Signed-off-by: glafond- <glafond-@student.42.fr>